+++ test/spec/propertiesPanel.spec.js
@@ -215,7 +215,34 @@ describe('Properties Panel - Data Group', function() {
 
     }));
 
+
+    it('should reject non-Object JSON', inject(async function(selection, elementRegistry) {
+
+      // given
+      const task = elementRegistry.get('empty');
+
+      await act(() => {
+        selection.select(task);
+      });
+
+      const entry = domQuery('.bio-properties-panel-entry', container);
+      const exampleInput = domQuery('textarea[name=exampleJson]', container);
+
+      // assume
+      await expectValid(entry);
+
+      // when
+      await act(() => {
+        changeInput(exampleInput, '1234');
+      });
+
+      // then
+      await expectInvalid(entry);
+
+    }));
+
   });
+
 });